mempool: Fix priority calculation during chain reorganizations
Fix assertion failure and potential undefined behavior when calculating
transaction priority during chain reorganizations where the spend height
is lower than the cached height.

Changes:
- Add GetCachedHeight() getter to CTxMemPoolEntry to allow callers to
  detect when cached priority data is stale due to chain rewinds
- Guard GetPriority() against unsigned integer underflow when
  spendheight < cachedHeight (legitimate during reorgs)
- Move priority calculation methods from coin_age_priority.cpp to their
  proper locations (txmempool.cpp, node/miner.cpp) to resolve circular
  dependency: kernel/mempool_entry -> policy/coin_age_priority
- Simplify coin_age_priority.cpp to contain only pure utility functions

This fixes a crash that could occur during block disconnection when
mempool entries had cached priority from a higher block height.
